# Optimized creation of bitmap by range