### SUMMARY
Optimized creation of bitmap by range via factory method `Util.bitmapOf()`. There is a trade-off between performance and  code is  duplication, but it can be ignored as piece of duplicated code is already duplicated several times in class `RoaringBitmap`.

| Benchmark | (offset) | (bytes) | Mode | Cnt | Score | Error | Units | 
| --- | --- | --- | --- | --- | --- | --- | --- |
| optimized | 0 | 10 | avgt | 5 | 31.364 | ±0.964 | ns/op |
| optimized | 0 | 100000 | avgt | 5 | 45.240 | ±3.783 | ns/op |
| optimized | 0 | 10000000 | avgt | 5 | 1663.824 | ±22.388 | ns/op |
| optimized | 100000 | 10 | avgt | 5 | 31.179 | ±0.358 | ns/op |
| optimized | 100000 | 100000 | avgt | 5 | 59.337 | ±9.726 | ns/op |
| optimized | 100000 | 10000000 | avgt | 5 | 1764.398 | ±329.763 | ns/op |
| original | 0 | 10 | avgt | 5 | 40.100 | ±1.431 | ns/op |
| original | 0 | 100000 | avgt | 5 | 65.978 | ±9.748 | ns/op |
| original | 0 | 10000000 | avgt | 5 |  3338.978 | ±379.935 | ns/op |
| original | 100000 | 10 | avgt | 5 | 40.919 | ±3.560 | ns/op |
| original | 100000 | 100000 | avgt | 5 | 84.212 | ±1.339 | ns/op |
| original | 100000 | 10000000 | avgt | 5 | 3295.778 | ±68.212 | ns/op |

Some tests added to verify correctness and improve code coverage.

### Automated Checks

- [x] I have run `./gradlew test` and made sure that my PR does not break any unit test.
- [x] I have run `./gradlew checkstyleMain` or the equivalent and corrected the formatting warnings reported.