@lemire 2023-11-17
Can you explain why it is faster? Didn't you just inline the original code...

```Java
     RoaringBitmap bitmap = new RoaringBitmap();
     bitmap.add(min, max);
```

???

I am concerned that the benchmark might be misleading.