@xtonik 2023-11-25
The main difference is that unnecessary copying of arrays in method [`RoaringArray#insertNewKeyValueAt(int, char, Container)`](https://github.com/RoaringBitmap/RoaringBitmap/blob/8dd4a5ec22531704ab834db66af89804c17fe37e/RoaringBitmap/src/main/java/org/roaringbitmap/RoaringArray.java#L794) is avoided using method [RoaringBitmap#append(char, Container)](https://github.com/RoaringBitmap/RoaringBitmap/blob/8dd4a5ec22531704ab834db66af89804c17fe37e/RoaringBitmap/src/main/java/org/roaringbitmap/RoaringBitmap.java#L3104) instead of. Above that resizing of RoaringArray is avoided also as it is precisely allocated beforehand.

Can you be more specific about misleading benchmark?