@lemire 2024-01-27
Thanks for the analysis. It makes sense.

Can you provide the same optimization to the buffer package? (E.g., producing a MutableRoaringBitmap).