@xtonik 2024-01-27
Version for buffered bitmap added.